Only perform featured-topic-related class modifications when enabled
These modifications are intricate and highly-coupled to core internals, so they are very likely to need updating over time as core changes. To reduce risk for the majority of users, we can move the conditionals outside the class modifications so that they only apply to people with the relevant settings enabled.

(this diff is mostly indentation changes)